estimate number of instances of a cache in describe_cache
If we know the number of logical CPUs in the socket, we can estimate the
number of instances of that cache within the socket.

For this, we rely on x2APIC or AMD leaf 0x80000008.

This currently doesn't work on my Ryzen 9 3950X in either WSL2 or in
Windows (privileged Hyper-V). I think Windows is stomping on the cache
values somehow, in particular the L3 number of threads sharing.

When I boot into bare metal, the L3 number of threads sharing is
correct.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>