feature: add more AMD features
Some of these aren't documented in the AMD manuals.

The Platform QoS ones are defined in the "AMD Technology Platform
Quality of Service" doc, but are not defined in the main CPU manuals.

The CLWB and INVPCID instructions are documented, but the CPUID bits for
them are not.

The REP MOVSB/STOSB ones are not documented as far as I could find, but
they are presumably there for Intel CPUID parity.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>